% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTht.R
\name{.rTht}
\alias{.rTht}
\title{Generate precision matrices for multivariate functional data}
\usage{
.rTht(p, id.diag, det.min = 0.5)
}
\description{
Generate precision matrices for multivariate functional data
}
