% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfdpo.R
\name{rmfdpo}
\alias{rmfdpo}
\title{Generate Partially Observed Multivariate Functional Data Under Partial Separability}
\usage{
rmfdpo(
  n,
  a = 0,
  b = 1,
  d = 60,
  w = 20,
  p,
  Sgm = NULL,
  basis = c("fourier", "bspline"),
  nbasis = 5,
  ...
)
}
\arguments{
\item{n}{the sample size.}

\item{a}{the lower bound of the full data domain. By default \code{a = 0}.}

\item{b}{the upper bound of the full data domain. By default \code{b = 1}.}

\item{d}{the number of full observed temporal points. By default \code{d = 60}.}

\item{w}{the number of missing temporal points. By default \code{w = 20}.}

\item{p}{the number of functional curves.}

\item{basis}{the type of basis that can be used. It is possible to choice \sQuote{\code{fourier}} or \sQuote{\code{bspline}} basis. By defaults \code{basis = "fourier"}.}

\item{nbasis}{the number of basis}
}
\value{
This function returns a list with the following objects:
\item{X}{ an array of the generated full observed multivariate functional data. By row the temporal points, by column the units, by the third dimension the functional curves.}
\item{Xpo}{an array of the generated partially observed multivariate functional data. By row the temporal points, by columns the units, by the third dimension the functional curves.}
\item{Xi}{an array of the true scores related to the full observed data. By row the units, by columns the functional curves, by the third dimension the basis.}
\item{U_mat}{ an array of the observed temporal grid. By row the temporal points, by columns the units, by the third dimension the functional covariates.}
\item{Tht}{ the true precision matrices under partial separability.}
\item{Sgm}{ the true var-cov matrices under partial separability.}
}
\description{
Generate Partially Observed Multivariate Functional Data Under Partial Separability assumption of the covariance operator.
}
\details{
This function generate a set of partially observed multivariate functional data under partial separability assumption of the covariance operator. Other argument of the function \link{rPar} can be specified to generate a more or less sparse variance-covariance matrix.
}
\examples{
set.seed(42)
rmfdpo(n = 100, p = 5)

}
