% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPar.R
\name{rPar}
\alias{rPar}
\title{Generate variance-covariance matrix for multivariate functional data}
\usage{
rPar(p, K, id.diag = 1, s1 = 3, s2 = -1.8)
}
\arguments{
\item{p}{the number of functional curves.}

\item{K}{the number of basis.}

\item{id.diag}{an argument that controls the sparsness of the matrices.}

\item{s1}{an scale factor.}

\item{s2}{a scale factor.}
}
\description{
This function generates a variance-covariance matrices and the corresponding precision matrices from a multivariate functional data under partial separability assumption of the covariance operator
}
